+++ Affirmation/app/src/main/res/values/colors.xml
