+++ Affirmation/gradle.properties
