+++ Affirmation/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Mon Nov 22 17:25:48 KST 2021
+distributionBase=GRADLE_USER_HOME
+distributionUrl=https\://services.gradle.org/distributions/gradle-7.0.2-bin.zip
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
