--- Affirmations/app/src/main/java/com/example/affirmations/MainActivity.kt
@@ -1,11 +1,0 @@
-package com.example.affirmations
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
