--- Affirmations/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Mon Nov 22 17:13:49 KST 2021
-distributionBase=GRADLE_USER_HOME
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.0.2-bin.zip
-distributionPath=wrapper/dists
-zipStorePath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
